Statistical inference for comparing empirical and expected ROCs. If CI=TRUE then also returns pointwise CIs

Statistical inference for comparing empirical and expected ROCs. If CI=TRUE then also returns pointwise CIs

mROC_inference(y, p, n_sim = 1e+05, CI = FALSE, aux = FALSE, fast = TRUE)

Arguments

y
vector of binary response values

p
vector of probabilities

n_sim
number of Monte Carlo simulations to calculate p-value

CI
optional. Whether confidence interval should be calculated for each point of mROC. Default is FALSE.

aux
aux optional. whether additional results (component-wise p-values etc) should be written in the package's aux variable. Default is FALSE.

fast
fast optional. Whether the fast code (C++) or slow code (R) should be called. Default is TRUE (R code will be slow unless the dataset is small)

Value

Returns an object of type mROC_inference containing the results of statistical inference for the mROC curve